requirements.txt
pandas
matplotlib
seaborn
scikit-learn
tensorflow-macos
joblib
